refactor(tests): cleanup tests
