+++ src/tsconfig.json
@@ -1,5 +1,6 @@
 {
     "compilerOptions": {
+        "noEmit": false,
         "outDir": "../dist",
     },
     "extends": "../tsconfig.json",
