+++ test/global.d.ts
@@ -1,0 +1,1 @@
+import '@types/jest';
