+++ tsconfig.json
@@ -7,13 +7,13 @@
         "forceConsistentCasingInFileNames": true,
         "module": "commonjs",
         "moduleResolution": "node",
+        "noEmit": true,
         "noEmitOnError": true,
         "noFallthroughCasesInSwitch": true,
         "noImplicitReturns": true,
         "noImplicitThis": false,
         "noUnusedLocals": true,
         "noUnusedParameters": true,
-        "outDir": "dist",
         "sourceMap": true,
         "strict": true,
         "target": "es6"
